refactor: extract InterestCalculator from Loan
Move the pure interest math (_compute_accrued_interest) and MoraStrategy
enum into a standalone InterestCalculator class. Loan delegates to
self._interest for all interest computation.
