+++ money_warp/loan/__init__.py
@@ -1,12 +1,14 @@
 """Loan module for personal loan modeling with flexible payment schedules."""
 
 from .installment import Installment
-from .loan import Loan, MoraStrategy
+from .interest_calculator import InterestCalculator, MoraStrategy
+from .loan import Loan
 from .settlement import AnticipationResult, Settlement, SettlementAllocation
 
 __all__ = [
     "AnticipationResult",
     "Installment",
+    "InterestCalculator",
     "Loan",
     "MoraStrategy",
     "Settlement",
